+++ src/main/drivers/accgyro_spi_mpu6500.c
@@ -164,33 +164,8 @@ static void mpu6500AccRead(int16_t *accData)
     accData[Z] = (int16_t)((buf[4] << 8) | buf[5]);
 }
 
-#ifdef COLIBRI_RACE
-static void mpu6500SpiInit(void)
-{
-    GPIO_InitTypeDef GPIO_InitStructure;
-
-    RCC_AHBPeriphClockCmd(MPU6500_CS_GPIO_CLK_PERIPHERAL, ENABLE);
-
-    GPIO_InitStructure.GPIO_Pin = MPU6500_CS_PIN;
-    GPIO_InitStructure.GPIO_Mode = GPIO_Mode_OUT;
-    GPIO_InitStructure.GPIO_Speed = GPIO_Speed_50MHz;
-    GPIO_InitStructure.GPIO_OType = GPIO_OType_PP;
-    GPIO_InitStructure.GPIO_PuPd = GPIO_PuPd_NOPULL;
-
-    GPIO_Init(MPU6500_CS_GPIO, &GPIO_InitStructure);
-
-    DISABLE_MPU6500;
-
-    spiSetDivisor(MPU6500_SPI, SPI_9MHZ_CLOCK_DIVIDER);
-}
-#endif
-
 static void mpu6500GyroInit(void)
 {
-#ifdef COLIBRI_RACE
-    mpu6500SpiInit();
-#endif
-
 #ifdef NAZE
     gpio_config_t gpio;
     // MPU_INT output on rev5 hardware (PC13). rev4 was on PB13, conflicts with SPI devices
